Refactor: move google & FB login buttons
move login buttons to containers/Login file and delete console logs.
