+++ src/containers/Login/Login.jsx
@@ -1,9 +1,38 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
+import firebase from 'firebase';
+import StyledFirebaseAuth from 'react-firebaseui/StyledFirebaseAuth';
+
+firebase.initializeApp({
+  apiKey: 'AIzaSyDXZEOi_czeNROEjfR9aWzEtVPEKec7rag',
+  authDomain: 'exhale-project-ff168.firebaseapp.com',
+});
 
 const Login = () => {
+  const [isSignedIn, setisSignedIn] = useState(false);
+  const uiConfig = {
+    signInFlow: 'popup',
+    signInOptions: [
+      firebase.auth.GoogleAuthProvider.PROVIDER_ID,
+      firebase.auth.FacebookAuthProvider.PROVIDER_ID,
+    ],
+  };
+  useEffect(() => {
+    firebase.auth().onAuthStateChanged((user) => {
+      setisSignedIn(!!user);
+    });
+  });
   return (
     <div>
-      <h1>Login</h1>
+      {isSignedIn ? (
+        <span>
+          <div className="text-red-600">Signed In!</div>
+          <button type="button" onClick={() => firebase.auth().signOut()}>
+            Sign out!
+          </button>
+        </span>
+      ) : (
+        <StyledFirebaseAuth uiConfig={uiConfig} firebaseAuth={firebase.auth()} />
+      )}
     </div>
   );
 };
